requirements.txt
absl-py==0.1.10
asn1crypto==0.24.0
autovizwidget==0.12.5
bleach==1.5.0
bokeh==0.12.14
boto3==1.5.34
botocore==1.8.48
certifi==2018.1.18
cffi==1.11.4
chardet==3.0.4
cryptography==2.1.4
cycler==0.10.0
decorator==4.2.1
docutils==0.14
entrypoints==0.2.3
environment-kernels==1.1.1
grpcio==1.8.3
h5py==2.7.1
hdijupyterutils==0.12.5
html5lib==0.9999999
idna==2.6
ipykernel==4.8.2
ipython==6.2.1
ipython-genutils==0.2.0
ipywidgets==7.1.2
jedi==0.11.1
Jinja2==2.10
jmespath==0.9.3
jsonschema==2.6.0
jupyter-client==5.2.2
jupyter-core==4.4.0
Keras==1.1.2
Markdown==2.6.11
MarkupSafe==1.0
matplotlib==2.1.2
mistune==0.8.3
nb-conda==2.2.1
nb-conda-kernels==2.1.0
nbconvert==5.3.1
nbformat==4.4.0
neural-image-analogies==0.1.2
notebook==5.4.0
numpy==1.14.1
opencv-contrib-python==3.4.0.12
packaging==16.8
pandas==0.22.0
pandocfilters==1.4.2
parso==0.1.1
pexpect==4.4.0
pickleshare==0.7.4
Pillow==5.0.0
plotly==2.4.0
prompt-toolkit==1.0.15
protobuf==3.5.1
psycopg2==2.7.4
ptyprocess==0.5.2
py4j==0.10.4
pycparser==2.18
pygal==2.4.0
Pygments==2.2.0
pykerberos==1.1.14
pyOpenSSL==17.5.0
pyparsing==2.2.0
PySocks==1.6.7
pyspark==2.2.1
python-dateutil==2.6.1
pytz==2018.3
PyYAML==3.12
pyzmq==17.0.0
requests==2.18.4
requests-kerberos==0.11.0
s3transfer==0.1.13
scikit-learn==0.19.1
scipy==1.0.0
Send2Trash==1.5.0
simplegeneric==0.8.1
six==1.11.0
sparkmagic==0.12.5
SQLAlchemy==1.2.3
tensorboard==1.0.0a4
tensorflow-gpu==1.5.0
tensorflow-tensorboard==1.5.1
terminado==0.8.1
testpath==0.3.1
Theano==1.0.1
tornado==4.5.3
traitlets==4.3.2
urllib3==1.22
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
widgetsnbextension==3.1.4
